+++ Pen/definitions.h
@@ -2,36 +2,45 @@
  * Definitions
  ********************************************************************/
 
-#define DEBUG false
-
 /*
  * Switch Bitmasks
  */
 #define ON_BM   0x08
 #define L_BM    0x10
 #define R_BM    0x20
-#define LR_BM    0x56
+#define LR_BM   0x56
 #define RR_BM   0x40
 #define OFF_BM  0x80
 
 /*
  * Pin Definitions
  */
-#define LED PIN_PA5 //PA5
-#define IR_LED_F  PIN_PB5
+#define LED       PIN_PC0 //PA5
+#define IR_LED_F  PIN_PA5
 #define IR_LED_R  PIN_PA4
+#define R_LED_F   PIN_PB4
+#define R_LED_R   PIN_PC3
 
-#define SW_L  PIN_PB2
-#define SW_RF PIN_PA6
-#define SW_RR PIN_PA2 
+#define SW_R      PIN_PB2
+#define SW_LF     PIN_PA6
+#define SW_LR     PIN_PB5 
+
+//SPI pins
+#define CS      PIN_PC1
+#define DATA    PIN_PA1
+#define SCK     PIN_PA3
+
+//Interrupt pins
+#define INT_IN PIN2_bm
 
 /*
  * PWM Settings
  */
-#define PWM_COUNT 121  //25 for 38khz, 29 for 33khz
-#define PWM_PRESCALER 0
-#define PWM_LOW PWM_COUNT/2  
-#define PWM_HIGH 255
+#define PWM_COUNT       131  //20 for 1 MHz, 131 for 5MHz
+#define PWM_PRESCALER   0
+#define PWM_LOW         PWM_COUNT/2  
+#define PWM_HIGH        0           //should be 255
+#define PWM_OFF         131           //should be 0
 
 /*
  * IR Timing Settings
@@ -45,3 +54,8 @@
 #define SIGROW  0x1100
 #define SER_X   11 
 #define SER_Y   12
+
+/*
+ * Misc
+ */
+#define TILT_THRESHOLD 15*TILT_ANGLE