+++ Sensor/data/variables.js
@@ -1,0 +1,2 @@
+const ipAddress = "";
+const wsPort = "";